+++ internal/app/core/ports/adapters/clients/clients.go
@@ -12,7 +12,7 @@ const (
 )
 
 type OrderLoyaltyInfo struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
